+++ C++_Player/CMakeLists.txt
@@ -6,12 +6,14 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${TORCH_CXX_FLAGS}")
 
 add_subdirectory(common)
 add_subdirectory(alpha_beta)
+add_subdirectory(human_play)
 add_executable(main main.cpp)
 
 target_link_libraries(main 
   "${TORCH_LIBRARIES}" 
   common
   alpha_beta
+  human_play
 )
 
 set_property(TARGET main PROPERTY CXX_STANDARD 14)
