+++ C++_Player/alpha_beta/CMakeLists.txt
@@ -4,8 +4,6 @@ add_library(alpha_beta
     alphabeta_board.h
     alphabeta.h
     alphabeta.cpp
-    humanplayer.h
-    humanplayer.cpp
 )
 
 target_include_directories(alpha_beta PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
