+++ project_summary.txt
@@ -15,3 +15,7 @@ TODO: 1) test the agent
 4/9/2022: Introducing transposition table to the alpha_beta pruning agent
 TODO: 1) start learning MCTS + Deep Learning algorithm
       2) further test the agent 
+
+5/9/2022: Refactor the code, seperate different players into different directory, and the common data structures into a different directory 
+TODO: 1) start learning the MCTS + Deep Learning algorithm
+      2) start learning socket programming for games
